chore: simplify types and remove dead code
