+++ app/routes/action.share.tsx
@@ -13,7 +13,7 @@ export const action: ActionFunction = async ({ context, request }) => {
     const shareId = nanoid();
     // expires in 1 week as UNIX second timestamp
     const expiration = new Date(Date.now() + 1000 * 60 * 60 * 24 * 1).getTime() / 1000;
-    await context.env?.SHARES?.put(shareId, theme, { expiration });
+    await context.env?.shares?.put(shareId, theme, { expiration });
     const shareUrl = new URL(request.url);
     shareUrl.pathname = `/themes/${shareId}`;
     return json({ shareUrl });
