+++ functions/[[path]].ts
@@ -4,5 +4,9 @@ import { createPagesFunctionHandler } from '@remix-run/cloudflare-pages';
 // @ts-ignore - the server build file is generated by `remix vite:build`
 // eslint-disable-next-line import/no-unresolved
 import * as build from '../build/server';
+import { getLoadContext } from '../load-context';
 
-export const onRequest = createPagesFunctionHandler({ build });
+export const onRequest = createPagesFunctionHandler({
+  build,
+  getLoadContext,
+});
