+++ package.json
@@ -6,13 +6,12 @@
   "sideEffects": false,
   "scripts": {
     "build": "remix vite:build",
-    "dev": "pnpm run --stream /^dev:/",
-    "dev:app": "remix vite:dev",
-    "start": "wrangler pages dev ./build/client",
-    "lint": "eslint --ignore-path .gitignore --cache --cache-location ./node_modules/.cache/eslint .",
-    "typecheck": "wrangler types && tsc",
-    "schema2ts": "pnpm json2ts app/state/themeFamily.json src/state/themeFamily.d.ts --unknownAny=false --style.singleQuote --additionalProperties=false",
     "build:validator": "npx tsx ./app/state/validator/buildValidator.ts",
+    "build:schema": "pnpm json2ts app/state/themeFamily.json src/state/themeFamily.d.ts --unknownAny=false --style.singleQuote --additionalProperties=false",
+    "dev": "remix vite:dev",
+    "dev:wrangler": "wrangler pages dev ./build/client",
+    "test:lint": "eslint --ignore-path .gitignore --cache --cache-location ./node_modules/.cache/eslint .",
+    "test:typecheck": "wrangler types && tsc",
     "sync:schemas": "npx tsx ./dev/fetchZedSchemas.ts"
   },
   "dependencies": {
