+++ worker-configuration.d.ts
@@ -1,0 +1,7 @@
+// Generated by Wrangler on Tue Mar 19 2024 18:24:16 GMT+0000 (Greenwich Mean Time)
+// by running `wrangler types`
+
+interface Env {
+	themes: KVNamespace;
+	shares: KVNamespace;
+}
